+++ lib/pin_keyboard_controller.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class PinKeyboardController {
+  VoidCallback? _resetCallback;
+
+  void addResetListener(VoidCallback listener) {
+    _resetCallback = listener;
+  }
+
+  void reset() {
+    if (_resetCallback != null) {
+      _resetCallback!();
+    }
+  }
+}